Revert "Added improve quotely and img editor"
This reverts commit 5a8a3c0130179045a27fdea4ac0a7753bf1cc5ae.
