--- Shikimori/modules/helper_funcs/quoteapi.py
@@ -1,186 +1,0 @@
-## Made By @Yash-Sharma-1807
-## telegram.dog/YASH_SHARMA_1807?start=papa
-
-import os 
-import base64
-from telethon.tl import types 
-from telethon.utils import get_display_name, get_peer_id
-import aiohttp
-from PIL import Image
-from aiohttp import ContentTypeError
-from Shikimori import DEV_USERS
-
-async def async_searcher(
-    url: str,
-    post: bool = None,
-    headers: dict = None,
-    params: dict = None,
-    json: dict = None,
-    data: dict = None,
-    ssl=None,
-    re_json: bool = False,
-    re_content: bool = False,
-    real: bool = False,
-    *args,
-    **kwargs,
-):
-    async with aiohttp.ClientSession(headers=headers) as client:
-        if post:
-            data = await client.post(
-                url, json=json, data=data, ssl=ssl, *args, **kwargs
-            )
-        else:
-            data = await client.get(url, params=params, ssl=ssl, *args, **kwargs)
-        if re_json:
-            return await data.json()
-        if re_content:
-            return await data.read()
-        if real:
-            return data
-        return await data.text()
-
-    
-class Quotly:
-    _API = "https://bot.lyo.su/quote/generate"
-    _entities = {
-        types.MessageEntityPhone: "phone_number",
-        types.MessageEntityMention: "mention",
-        types.MessageEntityBold: "bold",
-        types.MessageEntityCashtag: "cashtag",
-        types.MessageEntityStrike: "strikethrough",
-        types.MessageEntityHashtag: "hashtag",
-        types.MessageEntityEmail: "email",
-        types.MessageEntityMentionName: "text_mention",
-        types.MessageEntityUnderline: "underline",
-        types.MessageEntityUrl: "url",
-        types.MessageEntityTextUrl: "text_link",
-        types.MessageEntityBotCommand: "bot_command",
-        types.MessageEntityCode: "code",
-        types.MessageEntityPre: "pre",
-    }
-
-    async def _format_quote(self, event, reply=None, sender=None, type_="private"):
-        async def telegraph(file_):
-            file = file_ + ".png"
-            Image.open(file_).save(file, "PNG")
-            files = {"file": open(file, "rb").read()}
-            uri = (
-                "https://telegra.ph"
-                + (
-                    await async_searcher(
-                        "https://telegra.ph/upload", post=True, data=files, re_json=True
-                    )
-                )[0]["src"]
-            )
-            os.remove(file)
-            os.remove(file_)
-            return uri
-
-        if reply:
-            reply = {
-                "name": get_display_name(reply.sender) or "Deleted Account",
-                "text": reply.raw_text,
-                "chatId": reply.chat_id,
-            }
-        else:
-            reply = {}
-        is_fwd = event.fwd_from
-        name = None
-        last_name = None
-        if sender and sender.id not in DEV_USERS:
-            id_ = get_peer_id(sender)
-            name = get_display_name(sender)
-        elif not is_fwd:
-            id_ = event.sender_id
-            sender = await event.get_sender()
-            name = get_display_name(sender)
-        else:
-            id_, sender = None, None
-            name = is_fwd.from_name
-            if is_fwd.from_id:
-                id_ = get_peer_id(is_fwd.from_id)
-                try:
-                    sender = await event.client.get_entity(id_)
-                    name = get_display_name(sender)
-                except ValueError:
-                    pass
-        if sender and hasattr(sender, "last_name"):
-            last_name = sender.last_name
-        entities = []
-        if event.entities:
-            for entity in event.entities:
-                if type(entity) in self._entities:
-                    enti_ = entity.to_dict()
-                    del enti_["_"]
-                    enti_["type"] = self._entities[type(entity)]
-                    entities.append(enti_)
-        message = {
-            "entities": entities,
-            "chatId": id_,
-            "avatar": True,
-            "from": {
-                "id": id_,
-                "first_name": (name or (sender.first_name if sender else None))
-                or "Deleted Account",
-                "last_name": last_name,
-                "username": sender.username if sender else None,
-                "language_code": "en",
-                "title": name,
-                "name": name or "Unknown",
-                "type": type_,
-            },
-            "text": event.raw_text,
-            "replyMessage": reply,
-        }
-        if event.document and event.document.thumbs:
-            file_ = await event.download_media(thumb=-1)
-            uri = await telegraph(file_)
-            message["media"] = {"url": uri}
-
-        return message
-
-    async def create_quotly(
-        self,
-        event,
-        url="https://qoute-api-akashpattnaik.koyeb.app/generate",
-        reply={},
-        bg=None,
-        sender=None,
-        OQAPI=True,
-        file_name="quote.webp",
-    ):
-        """Create quotely's quote."""
-        if not isinstance(event, list):
-            event = [event]
-        if OQAPI:
-            url = Quotly._API
-        if not bg:
-            bg = "#1b1429"
-        content = {
-            "type": "quote",
-            "format": "webp",
-            "backgroundColor": bg,
-            "width": 512,
-            "height": 768,
-            "scale": 2,
-            "messages": [
-                await self._format_quote(message, reply=reply, sender=sender)
-                for message in event
-            ],
-        }
-        try:
-            request = await async_searcher(url, post=True, json=content, re_json=True)
-        except ContentTypeError as er:
-            if url != self._API:
-                return await self.create_quotly(
-                    self._API, post=True, json=content, re_json=True
-                )
-            raise er
-        if request.get("ok"):
-            with open(file_name, "wb") as file:
-                image = base64.decodebytes(request["result"]["image"].encode("utf-8"))
-                file.write(image)
-            return file_name
-        raise Exception(str(request))
-
-quotely = Quotly()
