+++ Shikimori/modules/imgeditor.py.txt
@@ -113,7 +113,7 @@ async def photo(client: pbot, message: Message):
                     ],
                 ]
             ),
-            reply_to_message_id=message.reply_to_message.message_id,
+            reply_to_message_id=message.reply_to_message_id,
         )
     except Exception as e:
         print("photomarkup error - " + str(e))
