+++ requirements.txt
@@ -20,7 +20,7 @@ geopy
 nltk
 psutil
 aiohttp
-Pillow
+Pillow>=9.0
 CurrencyConverter
 googletrans
 google-trans-new
@@ -90,4 +90,3 @@ pytube
 yt_dlp
 pytimeparse
 youtube_dl
-numpy
